backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
av==12.3.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
ctranslate2==4.6.0
distro==1.9.0
faiss-cpu==1.8.0.post1
fastapi==0.115.2
faster-whisper==1.0.3
filelock==3.19.1
flatbuffers==25.9.23
fsspec==2025.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
huggingface-hub==0.35.3
humanfriendly==10.0
idna==3.10
jiter==0.11.0
mpmath==1.3.0
numpy==1.26.4
onnxruntime==1.18.1
openai==1.46.0
packaging==25.0
protobuf==6.32.1
pycparser==2.23
pydantic==2.9.2
pydantic_core==2.23.4
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pyreadline3==3.5.4
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
setuptools==80.9.0
sniffio==1.3.1
soundfile==0.12.1
starlette==0.40.0
sympy==1.14.0
tiktoken==0.7.0
tokenizers==0.22.1
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.30.6
wheel==0.45.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9


